moving method to another file, as it is related to printer struct now
